Import NodeAddress re-ordering code from kubelet
When running with a legacy cloud provider, kubelet reorders
nodeaddresses to honour any --node-ip option it was passed. When running
with an external cloud provider this re-ordering must be done in the
cloud provider. To facilitate code-reuse we move the existing code from
kubelet into component-helpers so it can be called by both components.